refactor(sale): centralize types
